fix:pathes; login-conditional-render
