+++ src/app/pages/Auth/components/Register/Register.tsx
@@ -1,41 +1,9 @@
 import AuthForm from 'components/Forms/AuthForm';
-import { IInputForm } from 'config/types';
 
-import useInput from 'hooks/useInput';
-import useSubmit, { DispatchFormEnum, RegThunkResponse } from 'hooks/useSubmit';
-import { validateEmail, validateName, validatePassword, validateSimilarityPass } from 'utils/validate';
+import useReg from 'hooks/auth/useReg';
 
 const Register = () => {
-  const userName = useInput('', validateName);
-  const email = useInput('', validateEmail);
-  const password = useInput('', validatePassword);
-  const confirmPassword = useInput('', validateSimilarityPass(password.value));
-
-  const { handleSubmit } = useSubmit<RegThunkResponse>(DispatchFormEnum.REG, {
-    userName: userName.value,
-    email: email.value,
-    password: password.value,
-    confirmPassword: confirmPassword.value,
-  }); // as RegThunkResponse
-
-  const regInputs: IInputForm[] = [
-    { ...userName, name: 'userName', placeholder: 'Type name...' },
-    { ...email, name: 'email', placeholder: 'Type email...' },
-    {
-      ...password,
-      type: 'password',
-      name: 'password',
-
-      placeholder: 'Type password...',
-    },
-    {
-      ...confirmPassword,
-      name: 'confirmPassword',
-
-      type: 'password',
-      placeholder: 'Confirm password...',
-    },
-  ];
+  const { regInputs, handleSubmit } = useReg();
 
   return (
     <AuthForm
